# api/public/docs/openapi.yml
openapi: 3.1.0
info:
  title: 'Constellation APIs'
  license:
    name: 'GNU AFFERO GENERAL PUBLIC LICENSE'
    url: 'https://www.gnu.org/licenses/agpl-3.0.en.html'
  version: '2.0'
servers:
  -
    url: 'http://localhost:8080/api'
    description: 'The local environment.'
paths:
...
  /v1/permissions:
    get:
      tags:
        - Permission
      summary: 'Get a listing of the permissions'
      description: 'Get all permissions.'
      operationId: getPermissions
      parameters:
        -
          $ref: '#/components/parameters/base-filter-per-page'
        -
          $ref: '#/components/parameters/base-filter-page'
        -
          $ref: '#/components/parameters/base-filter-sort'
        -
          $ref: '#/components/parameters/base-filter-q'
        -
          name: 'filter[id]'
          in: query
          description: 'Filter by permission id.'
          schema:
            type: integer
        -
          name: 'filter[name]'
          in: query
          description: 'Filter by permission name.'
          schema:
            type: integer
      responses:
        '200':
          description: 'Permissions list'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  data: { type: array, items: { $ref: '#/components/schemas/resource-permission' } }
                  message: { type: string }
                  total: { type: integer }
                type: object
  /v1/roles:
    get:
      tags:
        - Role
      summary: 'Get a listing of the roles'
      description: 'Get all roles.'
      operationId: getRoles
      parameters:
        -
          $ref: '#/components/parameters/base-filter-per-page'
        -
          $ref: '#/components/parameters/base-filter-page'
        -
          $ref: '#/components/parameters/base-filter-sort'
        -
          $ref: '#/components/parameters/base-filter-q'
        -
          name: 'filter[id]'
          in: query
          description: 'Filter by role id.'
          schema:
            type: integer
        -
          name: 'filter[name]'
          in: query
          description: 'Filter by role name.'
          schema:
            type: integer
      responses:
        '200':
          description: 'Roles list'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  data: { type: array, items: { $ref: '#/components/schemas/resource-role' } }
                  message: { type: string }
                  total: { type: integer }
                type: object
    post:
      tags:
        - Role
      summary: 'Store an role'
      description: 'Store an role.'
      operationId: storeRole
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/request-create-role'
      responses:
        '200':
          description: 'Created role data.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/resource-role' }
                  message: { type: string }
                type: object
        '422':
          description: 'Invalid inputs.'
          content:
            application/json:
              schema:
                properties:
                  errors: { properties: { '{name-of-invalid-field}': { type: array, items: { description: 'Error message', type: string } } }, type: object }
                  message: { type: string }
                type: object
  '/v1/roles/{id}':
    get:
      tags:
        - Role
      summary: 'Display the specified role'
      description: 'Get an role.'
      operationId: showRole
      parameters:
        -
          name: id
          in: path
          description: 'id of the role'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Role detail'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  data: { type: array, items: { $ref: '#/components/schemas/resource-role' } }
                  message: { type: string }
                type: object
        '404':
          description: 'Resource not found'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string }
                  exception: { type: string }
                  file: { type: string }
                  line: { type: integer }
                  trace: { type: array, items: { properties: { file: { type: string }, line: { type: integer }, function: { type: string }, class: { type: string }, type: { type: string } }, type: object } }
                type: object
    put:
      tags:
        - Role
      summary: 'Update an role'
      description: 'Update an role.'
      operationId: updateRole
      parameters:
        -
          name: id
          in: path
          description: 'id of the role'
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/request-create-role'
      responses:
        '200':
          description: 'Updated role data.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/resource-role' }
                  message: { type: string }
                type: object
        '422':
          description: 'Invalid inputs.'
          content:
            application/json:
              schema:
                properties:
                  errors: { properties: { '{name-of-invalid-field}': { type: array, items: { description: 'Error message', type: string } } }, type: object }
                  message: { type: string }
                type: object
        '404':
          description: 'Resource not found'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string }
                  exception: { type: string }
                  file: { type: string }
                  line: { type: integer }
                  trace: { type: array, items: { properties: { file: { type: string }, line: { type: integer }, function: { type: string }, class: { type: string }, type: { type: string } }, type: object } }
                type: object
    delete:
      tags:
        - Role
      summary: 'Delete an role'
      description: 'Remove the specified role from storage.'
      operationId: deleteRole
      parameters:
        -
          name: id
          in: path
          description: 'id of the role'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Role deleted.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  message: { type: string }
                type: object
        '404':
          description: 'Role not found.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  message: { type: string }
                type: object
...
components:
  schemas:
...
    request-create-role:
      title: 'Create role'
      description: 'Create role request'
      required:
        - name
        - permissions
      properties:
        name:
          description: Name
          type: string
        team_id:
          description: 'Team id'
          type: array
          items:
            type: integer
      type: object
...
    resource-permission:
      title: Permission
      description: 'Permission resource'
      properties:
        id:
          description: Id
          type: integer
        name:
          description: Name
          type: string
        created_at:
          description: 'Creation date'
          type: string
          format: date-time
        updated_at:
          description: 'Last update date'
          type: string
          format: date-time
      type: object
    resource-role:
      title: Role
      description: 'Role resource'
      properties:
        id:
          description: Id
          type: integer
        name:
          description: Name
          type: string
        permissions:
          description: permissions
          type: array
          items:
            type: integer
        created_at:
          description: 'Creation date'
          type: string
          format: date-time
        updated_at:
          description: 'Last update date'
          type: string
          format: date-time
      type: object
...
  parameters:
    base-filter-per-page:
      name: perPage
      in: query
      description: 'Number of items displayed by page.'
      schema:
        type: integer
        default: 100
        enum:
          - 5
          - 10
          - 25
          - 100
          - 1000
    base-filter-page:
      name: page
      in: query
      description: 'Current page.'
      schema:
        type: integer
    base-filter-sort:
      name: sort
      in: query
      description: 'Field used to sort the result.'
      schema:
        type: string
      examples:
        string:
          summary: '-id'
          value: '-id'
    base-filter-q:
      name: q
      in: query
      description: 'Query string'
      required: false
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Basic Auth'
      scheme: bearer
security:
  -
    bearerAuth: []